Conditions can now be added to an Option node. Added appropriate validation. Added serialisation/deserialisation. Organised code into multiple files.
