+++ DialogueEditor/Assets/DialogueEditor/Assets/Scripts/SerializableCondition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0912dec35a0e7f446b89119b48718c76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
