+++ DialogueEditor/Assets/DialogueEditor/Assets/Scripts/SerializableNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c81c1b4bd15ae54c96ff25409737277
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
