+++ DialogueEditor/Assets/DialogueEditor/Assets/Scripts/SerializableParameter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe24a1a08deb13f46a25f672af773e33
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
