+++ DialogueEditor/Assets/DialogueEditor/Assets/Scripts/SerializeableConversation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf604bea2524244438cb8f7c28b2da3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
